requirements.txt
mysql-connector-python==8.3.0
bcrypt==4.1.3
python-dotenv==1.0.1